extract GroupedSubsetBase.h
